Add workflow: Publish to guide.ubuntu-mate.org
New commits to this repository will now automatically publish on
guide.ubuntu-mate.org. Up to now, this had to be manually invoked on the
guide.ubuntu-mate.org repository.

This is a copy of the guide.ubuntu-mate.org workflow, but adapted to
work in the context of this repository. This is to keep it simple
without webhooks. Plus, there is not much (infrastructure wise) that
should cause this to break.